api!: widen machine_id and node_id types from i32 to i64
BREAKING CHANGE: Change all machine_id and node_id parameters, fields, and
return types from `i32` to `i64` to support bit layouts that allocate 31
or more bits to either field.

**Why i32 was insufficient**

`max_machine_id()` was computed as `(1i32 << bits) - 1`. A 31-bit shift
shifts into the sign bit, which:
- Panics in debug mode (overflow check)
- Wraps to i32::MIN in release mode

This made any layout with 31 bits for machine_id or node_id (e.g.,
`BitLayout::new(16, 31, 0, 16)`) crash despite passing sum-to-63 validation.

i64 provides 63 bits of headroom — since machine_id_bits + node_id_bits can
never exceed 63, neither shift can overflow.

**Types Changed**

| Location | From (i32) | To (i64) |
|----------|-----------|----------|
| `SnowflakeError::MachineIdOutOfRange.given` | i32 | i64 |
| `SnowflakeError::MachineIdOutOfRange.max` | i32 | i64 |
| `SnowflakeError::NodeIdOutOfRange.given` | i32 | i64 |
| `SnowflakeError::NodeIdOutOfRange.max` | i32 | i64 |
| `SnowflakeError::NodeIdExhausted.max` | i32 | i64 |
| `SnowflakeComponents::machine_id` | i32 | i64 |
| `SnowflakeComponents::node_id` | i32 | i64 |
| `SnowflakeIdGenerator::machine_id` (field) | i32 | i64 |
| `SnowflakeIdGenerator::node_id` (field) | i32 | i64 |
| `SnowflakeIdGenerator::machine_id()` (getter) | i32 | i64 |
| `SnowflakeIdGenerator::node_id()` (getter) | i32 | i64 |
| All constructor parameters (`new`, `with_epoch`, `with_layout`, `with_layout_and_epoch`) | i32 | i64 |
| `SnowflakeIdBucket` constructor parameters | i32 | i64 |
| `global::init()` and `init_with_epoch()` machine_id | i32 | i64 |
| `BitLayout::max_machine_id()` and `max_node_id()` return | i32 | i64 |
| `MACHINE_ID` static (AtomicI32 → AtomicI64) | i32 | i64 |
| `NODE_COUNTER` static (AtomicI32 → AtomicI64) | i32 | i64 |

**Simplified Code**
- `assemble()` no longer needs `as i64` casts for machine_id and node_id

**Why This Is Breaking**
- External code passing i32 values will get type mismatches
- Getters now return i64 instead of i32
- Error match patterns expecting i32 fields will fail to compile

+++ src/bucket.rs
@@ -22,14 +22,14 @@ pub struct SnowflakeIdBucket {
 
 impl SnowflakeIdBucket {
     /// Constructs a bucket using the UNIX epoch and default [`BitLayout`].
-    pub fn new(machine_id: i32, node_id: i32) -> Result<Self, SnowflakeError> {
+    pub fn new(machine_id: i64, node_id: i64) -> Result<Self, SnowflakeError> {
         Self::with_epoch(machine_id, node_id, UNIX_EPOCH)
     }
 
     /// Constructs a bucket using a custom epoch and default [`BitLayout`].
     pub fn with_epoch(
-        machine_id: i32,
-        node_id: i32,
+        machine_id: i64,
+        node_id: i64,
         epoch: SystemTime,
     ) -> Result<Self, SnowflakeError> {
         let generator = SnowflakeIdGenerator::with_epoch(machine_id, node_id, epoch)?;
@@ -41,8 +41,8 @@ impl SnowflakeIdBucket {
 
     /// Constructs a bucket with a fully custom [`BitLayout`] and epoch.
     pub fn with_layout_and_epoch(
-        machine_id: i32,
-        node_id: i32,
+        machine_id: i64,
+        node_id: i64,
         layout: BitLayout,
         epoch: SystemTime,
     ) -> Result<Self, SnowflakeError> {
